Disconnect 'path' from 'name' fileset attribute
Disconnect 'path' from 'name' attribute for files and filesets.
Previously, the 'path' denoted the dir where the fileset or file
resided, and 'name' would be appended to it as the file's name, or the
fileset's root directory. This patch amends 'path' attribute resolution
so that 'name' is not part of it, and is only used as an identifier.
This is consistent with the way 'name' is used is general.